Redesign: help text before inputs in forms (#11570)
* Fix comments in FormBuilder

* Be consistent on the usage of the # notation for referring to the instance methods

* Add shared example for checking the help_text in form fields

* Change help_text order for #collection_check_boxes

* Change help_text order for #collection_radio_buttons

* Change help_text order for #data_picker

* Change help_text order for #check_box

* Change help_text order for private method field_with_validations

As this is used for almost all the classic fields (i.e. `text_input`),
then we test this change indirectly by checking it out in the different
examples.

* Fix disposition of the input character counter

* Fix CSS for help-text

In some cases (for instance, in numeric fields) it was showing with text bold,
as it was using the CSS from the label.

* Implement help_text for #scopes_picker_filter

* Use _html suffix instead of calling the html_safe method

* Implement help_text for #editor

* Change (again) order for help_text in #check_box

Suggested by code review. In this particular input, placing the
help_text before the label doesn't make any sense.

* Add help_text spec for #geocoding_field

* Respect help_text option in form builders

Mostly in the admin panel, there were lots of places were we where
hardcoding the `help-text` like this:

```erb
<%= form.date_field :duration %>
<p class="help-text"><%== t(".duration_help") %></p>
```

This doesn't allow respecting the actual FormBuilder help-text, so it's
better to change it to use the help_text as an option:

```erb
<%= form.date_field :duration, help_text: t(".duration_help") %>
```

This way if tomorrow we introduce changes in the help_text (for
instance, using another class), then it's cleaner.

* Fix i18n-tasks offenses

* Fix erblint offenses

* Fix specs

* Add spec for checking that help text text is only rendered once

* Fix help text rendered as an attribute in #check_box

* Undo help-text in label

* Add spec for #select and #select with help-text

* Fix bug with help-text in meetings' #select inputs

* Add missing i18n string in meeting forms

* Use _html suffix for `cta_path_help_html`

* Undo help-text for templates' previews

* Fix bug with help-text in #editor input

* Implement change for checking that the help text text is rendered once in other specs

* Fix spec related to the change of 'Log in'/'Log out'