Make the `pwa_chrome` Capybara driver not open pages
While working in the `./decidim-core/spec/system/account_spec.rb`,
I've detected that we have not enabled the headless option, meaning that
it is opening pages in a visibly browser. This is a bit weird because
it's the only driver that's doing that.

This commit disables that behavior so its more consistent.