Apply flash styles to Announcements (#11708)
* Apply HTML from flashes to announcements

* Inline CSS classes to HTML callout to fix usage while the flash is actually an announcement

* Migrate hardcoded callouts to AnnouncementCell

* Fix flash styles

* Remove `.flex-col`, as when there are links it breaks. For instance,
  in the decidim-elections Bulletin Board not configured message in the
  Admin panel.
* Replace `.font-bold` with `.font-semibold`, reducing the boldness
* Add underline to links, so users can know that they're clickable

* Make the flash alerts manually in the cases where it can't be created with AnnouncementCell

There are some cases where we can't use the Cell, as there are links and
there's a sanititzation process.

On these cases we add a comment and style them manually.

* Fix bug with size of SVG icons in Admin

This rule was overriding all of the SVG, so we relax it a bit.

* Fix columns on flash when there's a title in the AnnouncementCell

* Fix `have_admin_callout` matcher with multiple flashes in the same page

* Fix specs

* Add full flash messages on 'has questionnaire' shared spec

* Improve `css_class` method when there isn't any `callout_class` argument

* Normalize i18n files with i18n-tasks

* Fix hidding when canceling email change in account#show

Until now the Javascript `hidden` function worked without issues, but now that we're
changing classes of this element, as the `flash` class has already a
display, it doesn't get hidden.

The culprit is:

> Web browsers may implement the hidden state using display: none, in which case the
> element will not participate in page layout. This also means that changing the value
> of the CSS display property on an element in the hidden state will override the state.
> For instance, elements styled display: block will be displayed despite the hidden
> attribute's presence.

https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/hidden#the_hidden_state

So, with a god old `display: none` we can fix this problem.

* Make the `pwa_chrome` Capybara driver not open pages

While working in the `./decidim-core/spec/system/account_spec.rb`,
I've detected that we have not enabled the headless option, meaning that
it is opening pages in a visibly browser. This is a bit weird because
it's the only driver that's doing that.

This commit disables that behavior so its more consistent.

* Revert "Make the pwa_chrome Capybara driver not open pages"

I left a comment on why I'm reverting this. Basically, the offline
navigation feature doesn't work with the headless browser.

* Change CSS class for callout

Suggested by code review

* Drop truncate support from AnnouncementCell

It was implemented in two places and it didn't read well, so we're
removing this altogether.