Remove unecessary method `check_current_user_can_visit_space`
This was only used in the :show action of the ConferencesController, and
as it's no longer used then we can remove it.

Also remove calls to and checks to `current_user_can_visit_space?`, as
this is only used for spaces that support the private/transparent
attributes. This is not the case in Conferences.