Remove `participatory_space_layout` helper method (#12021)
* Remove `participatory_space_layout` helper method

* Inline follow buttons on the necessary pages

Mind that we're only moving it in the :show action, and not in the
others (like the :index in the AssemblyMembersController) as it doesn't
make much sense to have it on those pages.

* Add enforce_permission_to to fix failing spec

This also brings consistency with other spaces show actions (such as
AssembliesController and ParticipatoryProcessesController)

* Make consistent the comments in the spaces ApplicationControllers

* Remove unecessary method `check_current_user_can_visit_space`

This was only used in the :show action of the ConferencesController, and
as it's no longer used then we can remove it.

Also remove calls to and checks to `current_user_can_visit_space?`, as
this is only used for spaces that support the private/transparent
attributes. This is not the case in Conferences.

* Convert docs to yardocs

* Migrate markup from layout to pages in participatory processes

* Migrate markup from layout to pages in assemblies

* Migrate markup from layout to pages in conferences

* Migrate markup from layout to pages in votings

In this case we're also adding the meta tags that weren't in this space.

* Remove unecessary param twitter_handler

It's already defined in the core meta tags. As we're not overriding
it, it doesn't make sense to have it in all these show pages.

* Fix a11y spec

* Fix markup in assemblies to use main correctly