Add custom configuration for non standard rspec filename
For some reason, rspec doesn't like that this particular file is called
"oauth_application_presenter_spec.rb", saying that it should be called
"o_auth_application_presenter_spec.rb".

The problem is that for consistency we should rename all the other
OAuth files. Also, it wasn't detected for other specs, like:

- ./decidim-system/spec/commands/decidim/system/create_oauth_application_spec.rb
- ./decidim-system/spec/commands/decidim/system/destroy_oauth_application_spec.rb
- ./decidim-system/spec/commands/decidim/system/update_oauth_application_spec.rb
- ./decidim-system/spec/forms/decidim/system/oauth_application_form_spec.rb

So, this commit changes the rspec configuration so this doesn't get
flagged for this particular rspec file.