Fix performance issue with attribute encryption/decryption (#12568)
* Fix performance issue with attribute encryption

Creating new instances of the `ActiveSupport::MessageEncryptor`
causes delay every time an attribute value is decrypted. We are
using exactly the same key and options for all attributes, so
we can cache this variable and by doing so highly improve the
performance.

* Remove unnecessary comment

* Fix the attribute encryptor specs

* Clear the shared `@cryptor` variable before and after specs

In order to ensure that no previous specs are interfering with
this spec.