Add votes count to comment caches
* Add votes count to comment caches

* Use size instead of count method

Co-authored-by: Alexandru Emil Lupu <contact@alecslupu.ro>

* Implement counter_cache for comment's votes

We need to handle the counter_cache manually as we have an scoped
association (i.e. with the lambda that does the where in the relation)

> has_many :up_votes, -> { where(weight: 1) }

So, for this reason we need to do handle the update with a callback in
CommentVote and can't use the reset_counters in the migration.

* Explicitly use the counter_cache attributes in comment cell

---------

Co-authored-by: Alexandru Emil Lupu <contact@alecslupu.ro>