Fix issues serving image variants directly through the storage service
The image variants need to be processed through the local
representation URLs when they are served the first time because
this is when these assets are processed for the first time.

These changes detect if the variant needs processing and serves
it through the local representation URL before the image has been
processed. If the image is already processed it is served directly
through the storage service itself.