Fix specs broken due to the changes in the asset URL generation
As the URLs are now changed every time they are generated, this
caused some specs to fail that assumed the URL of the asset is
matches always the format in which it was first generated.

The asset URL includes the expiry information and the expiry
information is different every time the URL is requested.