Roll our own counter_cache for attachments_count
The problem is basically that in the case of the DummyResource class, we
don't have a way of disabling this attribute, as it isn't a valid
attribute in the table and that's what expects counter_cache.

The solution is to have a more granular control of how the counter_cache
is updated by manually handling the increment/decrement of the fields.