Implement .table-stacked layout for responsive
I tried to make the action buttons dropdown for admin tables compatible
with our current solution for responsiveness in these tables: .table-scroll.

This uses the CSS property for overflow-x auto, and it doesn't play
nice with having an element outside of this scroll. I don't want to
introduce JS for getting out of this table neither.

So, my proposal is to implement a table stacked layout for the admin tables.

This means that we need to do a couple of changes to the tables:

1. Use the .table-stacked class instead of .table-scroll
2. Add the data-labels to the name of `td` so it can be shown in the responsive
3. For the action buttons disposition, wrap them in `<div class="inline-block relative">`